FastAPI/events-public/src/requirements.txt
fastapi
uvicorn
google-cloud-firestore
numpy
pydantic